Use tXXDController everywhere (better timing).  Also fix a bug in the previous commit for cmd.is_activate
